much OCD.. such wow..  very shiny....
I tried to OCD up the orginization of my assets.
There is a word document with my new sources now included.
there is a notepad document with the pixel width and height of each
frame for spritesheets, so it makes it faster to program them in without
having to check the pixels.